Include custom linux builds and custom void-packages fork.
This includes custom linux builds and it uses my own void-packages fork,
as that is a bit easier to maintain for myself than having something like
a set of patches here, as such most of what is located is pretty simple.

It also includes a new variable (XBPS_PACKAGES_REPO) to point to my custom
repo, incase it changes in the future.

I've also included documentation on everything in void-packages (aside from Discord)
to explain why I've gone and built it myself.