Update BASH configurations.
Configs have been update as necessary, with changes moved as needed.

The singular line setting path has in bash_profile is maintained because
it is only to graphical programs, and if we have a non-login shell of
the account we shouldn't have those accessible to us if not previously
logged in.